separate form (and client code) from profile page
